Split off executor into it's own file
