+++ enterprise/internal/campaigns/reconciler_test.go
@@ -5,14 +5,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/pkg/errors"
-
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns/store"
 	ct "github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns/testing"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
-	"github.com/sourcegraph/sourcegraph/internal/db"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/github"
@@ -205,105 +202,4 @@ func buildGithubPR(now time.Time, externalState campaigns.ChangesetExternalState
 	}
 
 	return pr
-}
-
-func TestDecorateChangesetBody(t *testing.T) {
-	db.Mocks.Namespaces.GetByID = func(ctx context.Context, org, user int32) (*db.Namespace, error) {
-		return &db.Namespace{Name: "my-user", User: user}, nil
-	}
-	defer func() { db.Mocks.Namespaces.GetByID = nil }()
-
-	internalClient = &mockInternalClient{externalURL: "https://sourcegraph.test"}
-	defer func() { internalClient = api.InternalClient }()
-
-	fs := &FakeStore{
-		GetCampaignMock: func(ctx context.Context, opts store.GetCampaignOpts) (*campaigns.Campaign, error) {
-			return &campaigns.Campaign{ID: 1234, Name: "reconciler-test-campaign"}, nil
-		},
-	}
-
-	cs := ct.BuildChangeset(ct.TestChangesetOpts{OwnedByCampaign: 1234})
-
-	body := "body"
-	rcs := &repos.Changeset{Body: body, Changeset: cs}
-	if err := decorateChangesetBody(context.Background(), fs, rcs); err != nil {
-		t.Errorf("unexpected non-nil error: %v", err)
-	}
-	if want := body + "\n\n[_Created by Sourcegraph campaign `my-user/reconciler-test-campaign`._](https://sourcegraph.test/users/my-user/campaigns/reconciler-test-campaign)"; rcs.Body != want {
-		t.Errorf("repos.Changeset body unexpectedly changed:\nhave=%q\nwant=%q", rcs.Body, want)
-	}
-}
-
-func TestCampaignURL(t *testing.T) {
-	ctx := context.Background()
-
-	t.Run("errors", func(t *testing.T) {
-		for name, tc := range map[string]*mockInternalClient{
-			"ExternalURL error": {err: errors.New("foo")},
-			"invalid URL":       {externalURL: "foo://:bar"},
-		} {
-			t.Run(name, func(t *testing.T) {
-				internalClient = tc
-				defer func() { internalClient = api.InternalClient }()
-
-				if _, err := campaignURL(ctx, nil, nil); err == nil {
-					t.Error("unexpected nil error")
-				}
-			})
-		}
-	})
-
-	t.Run("success", func(t *testing.T) {
-		internalClient = &mockInternalClient{externalURL: "https://sourcegraph.test"}
-		defer func() { internalClient = api.InternalClient }()
-
-		url, err := campaignURL(
-			ctx,
-			&db.Namespace{Name: "foo", Organization: 123},
-			&campaigns.Campaign{Name: "bar"},
-		)
-		if err != nil {
-			t.Errorf("unexpected non-nil error: %v", err)
-		}
-		if want := "https://sourcegraph.test/organizations/foo/campaigns/bar"; url != want {
-			t.Errorf("unexpected URL: have=%q want=%q", url, want)
-		}
-	})
-}
-
-func TestNamespaceURL(t *testing.T) {
-	t.Parallel()
-
-	for name, tc := range map[string]struct {
-		ns   *db.Namespace
-		want string
-	}{
-		"user": {
-			ns:   &db.Namespace{User: 123, Name: "user"},
-			want: "/users/user",
-		},
-		"org": {
-			ns:   &db.Namespace{Organization: 123, Name: "org"},
-			want: "/organizations/org",
-		},
-		"neither": {
-			ns:   &db.Namespace{Name: "user"},
-			want: "/users/user",
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			if have := namespaceURL(tc.ns); have != tc.want {
-				t.Errorf("unexpected URL: have=%q want=%q", have, tc.want)
-			}
-		})
-	}
-}
-
-type mockInternalClient struct {
-	externalURL string
-	err         error
-}
-
-func (c *mockInternalClient) ExternalURL(ctx context.Context) (string, error) {
-	return c.externalURL, c.err
 }
